+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added liability support
 - Better training for intelligence model
 - Added new agentic workflows
+- Added multimodel support for intelligence
 - Migrated from Redis to MongoDB for token storage due to frequent redis disconnection issue
 
 ## v1.0